fix(irq): pass full APIC IDs to IOAPIC routing
Preserve the widened IOMMU routing contract at serial and PS/2 call sites.

+++ crates/drivers/drv-ps2-keyboard/src/imp/irq.rs
@@ -99,7 +99,7 @@ unsafe fn install_line(
     let bsp_apic = BSP_APIC_ID.load(Ordering::Acquire) as u8;
     // SAFETY: a handler owns `vec`, the I/O-APIC window is mapped, and this
     // probe owns the line until the matching published pin is taken.
-    if !unsafe { arch_irq::program_x86_ioapic(pin, vec, bsp_apic, level, active_low) } {
+    if !unsafe { arch_irq::program_x86_ioapic(pin, vec, u32::from(bsp_apic), level, active_low) } {
         let _ = arch_irq::free_x86_vector(vec);
         return false;
     }
